requirements.txt
appdirs
python-ldraw
PyYaml